Divide existing code into modules
