+++ js/setup.js
@@ -1,0 +1,4 @@
+const COMMENT_MAX_LENGTH = 140;
+const DESCRIPTIONS_COUNT = 25;
+
+export { COMMENT_MAX_LENGTH, DESCRIPTIONS_COUNT };
